# public/api-docs/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/beneficio": {
      "get": {
        "description": "Retorna uma lista paginada de todos os tipos de benefícios cadastrados no sistema.",
        "operationId": "Beneficio_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Página atual",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Itens por página",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Termo de busca (nome)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ativo",
            "required": false,
            "in": "query",
            "description": "Filtro por status",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de benefícios retornada com sucesso",
            "content": {
              "application/json": {
                "example": ""
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar tipos de benefícios",
        "tags": [
          "Benefícios"
        ]
      },
      "post": {
        "description": "Cria um novo tipo de benefício no sistema.",
        "operationId": "Beneficio_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTipoBeneficioDto"
              },
              "examples": {
                "Auxílio Emergencial": {
                  "value": ""
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Benefício criado com sucesso",
            "content": {
              "application/json": {
                "example": ""
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "example": ""
              }
            }
          },
          "409": {
            "description": "Nome já em uso",
            "content": {
              "application/json": {
                "example": ""
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar novo tipo de benefício",
        "tags": [
          "Benefícios"
        ]
      }
    },
    "/api/v1/beneficio/{id}": {
      "get": {
        "description": "Retorna os detalhes completos de um tipo de benefício específico.",
        "operationId": "Beneficio_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Benefício encontrado com sucesso",
            "content": {
              "application/json": {
                "example": ""
              }
            }
          },
          "404": {
            "description": "Benefício não encontrado",
            "content": {
              "application/json": {
                "example": ""
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter detalhes de um benefício",
        "tags": [
          "Benefícios"
        ]
      },
      "put": {
        "description": "Atualiza os dados de um tipo de benefício existente.",
        "operationId": "Beneficio_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTipoBeneficioDto"
              },
              "examples": {
                "Atualização de Benefício": {
                  "value": ""
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Benefício atualizado com sucesso",
            "content": {
              "application/json": {
                "example": ""
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "example": ""
              }
            }
          },
          "404": {
            "description": "Benefício não encontrado",
            "content": {
              "application/json": {
                "example": ""
              }
            }
          },
          "409": {
            "description": "Nome já em uso",
            "content": {
              "application/json": {
                "example": ""
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar tipo de benefício existente",
        "tags": [
          "Benefícios"
        ]
      }
    },
    "/api/v1/beneficio/{id}/requisitos": {
      "get": {
        "operationId": "Beneficio_findRequisitos",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de requisitos retornada com sucesso"
          },
          "404": {
            "description": "Benefício não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar requisitos documentais",
        "tags": [
          "Benefícios"
        ]
      },
      "post": {
        "operationId": "Beneficio_addRequisito",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequisitoDocumentoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Requisito adicionado com sucesso"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "404": {
            "description": "Benefício não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Adicionar requisito documental",
        "tags": [
          "Benefícios"
        ]
      }
    },
    "/api/v1/beneficio/{id}/fluxo": {
      "put": {
        "operationId": "Beneficio_configurarFluxo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigurarFluxoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Fluxo configurado com sucesso"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "404": {
            "description": "Benefício não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Configurar fluxo de aprovação",
        "tags": [
          "Benefícios"
        ]
      }
    },
...
  },
  "info": {
    "title": "API de Gestão de Benefícios Eventuais",
...
    "version": "1.0",
    "contact": {
      "name": "Equipe SEMTAS",
      "url": "https://www.natal.rn.gov.br/semtas",
      "email": "suporte@semtas.natal.rn.gov.br"
    },
    "license": {
      "name": "Uso Interno",
      "url": "https://www.natal.rn.gov.br/semtas/licenca"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Insira o token JWT",
        "in": "header"
      }
    },
    "schemas": {
...
      "CriteriosElegibilidadeDto": {
        "type": "object",
        "properties": {
          "idade_minima": {
            "type": "number",
            "description": "Idade mínima para elegibilidade"
          },
          "idade_maxima": {
            "type": "number",
            "description": "Idade máxima para elegibilidade"
          },
          "renda_maxima": {
            "type": "number",
            "description": "Renda máxima per capita para elegibilidade"
          },
          "tempo_minimo_residencia": {
            "type": "number",
            "description": "Tempo mínimo de residência no município (em meses)"
          },
          "outros": {
            "description": "Outros critérios de elegibilidade",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateTipoBeneficioDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "description": "Nome do tipo de benefício",
            "example": "Auxílio Moradia"
          },
          "descricao": {
            "type": "string",
            "description": "Descrição detalhada do benefício"
          },
          "periodicidade": {
            "type": "string",
            "description": "Periodicidade do benefício",
            "enum": [
              "unico",
              "mensal",
              "trimestral",
              "semestral",
              "anual"
            ],
            "example": "mensal"
          },
          "base_juridica": {
            "type": "string",
            "description": "Base jurídica que fundamenta o benefício"
          },
          "valor": {
            "type": "number",
            "description": "Valor do benefício",
            "example": 400
          },
          "ativo": {
            "type": "boolean",
            "description": "Status do benefício",
            "default": true
          },
          "criterios_elegibilidade": {
            "description": "Critérios de elegibilidade para o benefício",
            "allOf": [
              {
                "$ref": "#/components/schemas/CriteriosElegibilidadeDto"
              }
            ]
          }
        },
        "required": [
          "nome",
          "descricao",
          "periodicidade",
          "base_juridica",
          "valor"
        ]
      },
      "UpdateTipoBeneficioDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "description": "Nome do tipo de benefício",
            "example": "Auxílio Moradia"
          },
          "descricao": {
            "type": "string",
            "description": "Descrição detalhada do benefício"
          },
          "periodicidade": {
            "type": "string",
            "description": "Periodicidade do benefício",
            "enum": [
              "unico",
              "mensal",
              "trimestral",
              "semestral",
              "anual"
            ],
            "example": "mensal"
          },
          "base_juridica": {
            "type": "string",
            "description": "Base jurídica que fundamenta o benefício"
          },
          "valor": {
            "type": "number",
            "description": "Valor do benefício",
            "example": 400
          },
          "ativo": {
            "type": "boolean",
            "description": "Status do benefício",
            "default": true
          },
          "criterios_elegibilidade": {
            "description": "Critérios de elegibilidade para o benefício",
            "allOf": [
              {
                "$ref": "#/components/schemas/CriteriosElegibilidadeDto"
              }
            ]
          }
        }
      },
      "CreateRequisitoDocumentoDto": {
        "type": "object",
        "properties": {
          "tipo_documento": {
            "type": "string",
            "description": "Tipo do documento requerido",
            "enum": [
              "rg",
              "cpf",
              "comprovante_residencia",
              "comprovante_renda",
              "certidao_nascimento",
              "declaracao_medica",
              "contrato_aluguel",
              "outro"
            ],
            "example": "comprovante_residencia"
          },
          "descricao": {
            "type": "string",
            "description": "Descrição do documento e suas especificações",
            "example": "Comprovante de residência dos últimos 3 meses (conta de água, luz ou telefone)"
          },
          "obrigatorio": {
            "type": "boolean",
            "description": "Indica se o documento é obrigatório para a solicitação",
            "default": true
          },
          "observacoes": {
            "type": "string",
            "description": "Observações adicionais sobre o documento",
            "example": "Caso não possua comprovante em seu nome, apresentar declaração do titular"
          }
        },
        "required": [
          "tipo_documento",
          "descricao",
          "obrigatorio"
        ]
      },
      "EtapaFluxoDto": {
        "type": "object",
        "properties": {
          "ordem": {
            "type": "number",
            "description": "Ordem da etapa no fluxo de aprovação",
            "example": 1
          },
          "nome": {
            "type": "string",
            "description": "Nome da etapa",
            "example": "Análise Técnica"
          },
          "descricao": {
            "type": "string",
            "description": "Descrição da etapa",
            "example": "Análise técnica da documentação e critérios de elegibilidade"
          },
          "tipo_aprovador": {
            "type": "string",
            "description": "Tipo de aprovador responsável pela etapa",
            "enum": [
              "tecnico",
              "coordenador_cras",
              "coordenador_creas",
              "gestor_unidade",
              "gestor_semtas",
              "comissao"
            ],
            "example": "tecnico"
          },
          "prazo_dias": {
            "type": "number",
            "description": "Prazo máximo em dias para conclusão da etapa",
            "example": 5
          }
        },
        "required": [
          "ordem",
          "nome",
          "descricao",
          "tipo_aprovador"
        ]
      },
      "ConfigurarFluxoDto": {
        "type": "object",
        "properties": {
          "descricao": {
            "type": "string",
            "description": "Descrição geral do fluxo de aprovação",
            "example": "Fluxo padrão para aprovação de benefícios eventuais"
          },
          "etapas": {
            "description": "Etapas do fluxo de aprovação",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EtapaFluxoDto"
            }
          }
        },
        "required": [
          "descricao",
          "etapas"
        ]
      },
...
    }
  }
}